Modularize touch input module.
